+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes qsf-integration project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-
+- add a possibility to stop a running pipeline
 
 ### Changed